fix(react-doctor): move OrlandoParksGallery data to its own module
React Doctor flagged only-export-components on the ALL_PARK_NAMES
export added in the last commit: a component file exporting
non-component values breaks Fast Refresh's ability to preserve state.

Moves FEATURED_PARKS, OTHER_PARK_GROUPS, TOTAL_PARKS_COUNT, and
ALL_PARK_NAMES into components/landings/orlando/orlandoParksData.ts,
matching the existing components/parques-brasil/constants.ts
convention — OrlandoParksGallery.tsx now only exports the component.
Updates tests/static-media-assets.test.ts, which read the component
file's source for image-hotlink checks, to also read the new data
file so it doesn't silently lose coverage of the moved image paths.
